+++ CHANGES.md
@@ -5,12 +5,23 @@
 
 ### Overview
 
+* Change "maximum timeout" setting meaning.
+* Add lock timeout user setting.
 * Bug fixes.
 
 ### Details
 
+* Change "maximum timeout" setting meaning.
+    * The "maximum timeout" user setting hasn't been used since the 0.7
+      release, so it now means the maximum SO socket time to live, which
+      allows the user to avoid a potential issue with the underlying
+      Perforce API. (#85)
+* Add lock timeout user setting.
+    * Allows the user to adjust how long the plugin waits for the connection
+      until the next one comes free.
+    * Useful for users that have a very slow connection to the server.
 * Bug fixes.
-    * ...
+    * *In progress*
 
 
 ## ::v0.8.5::
@@ -34,6 +45,10 @@
     * Fixed the configuration panel width - an outer scroll pane confused the
       tab layout.  The blank text fields shouldn't scroll offscreen anymore.
     * Fixed the SSL Fingerprint text field to correctly show the value.
+    * Went back to actually using the "reconnect with each request" setting.
+      Before, this setting was ignored and all connections were reconnected
+      before being used.  Users with slow connections should see a performance
+      boost with this disabled.
 
 
 ## ::v0.8.4::